fix: don't match a trailing period after the path
Fixes #17.

This also extends `trailingPeriod: false` (the default) to exclude
matching a trailing question mark (?) or exclamation mark (!).

That is if you want "Check out example.com/project." to NOT match the
trailing period, you typically also want it to NOT match the trailing
exclamation mark in "Check out example.com/project!".
Likewise for "Have you seen example.com/project?".

I guess technically `trailingPeriod` should be renamed
`trailingPunctuation` but maybe it's fine as-is?